Refactoring PCTL analysis - moved to separate class, takes transition system instead of model
