+++ Parsing/ParseBCSL.py
@@ -42,6 +42,8 @@ def load_TS_from_json(json_file: str) -> TransitionSystem:
         ts.states_encoding = {MemorylessState(np.array(eval(data['nodes'][node_id]))): int(node_id) for node_id in data['nodes']}
         ts.edges = {edge_from_dict(edge) for edge in data['edges']}
         ts.init = data['initial']
+        if 'parameters' in data:
+            ts.params = data['parameters']
 
         ts.unprocessed = {MemorylessState(np.array(eval(state))) for state in data.get('unprocessed', list())}
         ts.processed = ts.states_encoding.keys() - ts.unprocessed
